Fix hello-world
